requirements.txt
Django
psycopg2-binary
requests
pandas
openpyxl